CC36: breadth first graph
