+++ javascript/data-structure/graph/edge.js
@@ -1,0 +1,10 @@
+'use strict';
+
+class Edge {
+  constructor(vertex, weight = 0) {
+    this.vertex = vertex;
+    this.weight = weight;
+  }
+}
+
+module.exports = Edge;
