+++ javascript/data-structure/graph/graph.js
@@ -1,19 +1,6 @@
 'use strict';
 
-const { Queue } = require('../stacksAndQueues/stacks-and-queues');
-
-class Vertex {
-  constructor(value) {
-    this.value = value;
-  }
-}
-
-class Edge {
-  constructor(vertex, weight = 0) {
-    this.vertex = vertex;
-    this.weight = weight;
-  }
-}
+const Edge = require('./edge');
 
 class Graph {
   constructor() {
@@ -26,7 +13,10 @@ class Graph {
   }
 
   addEdge(startVertex, endVertex, weight) {
-    if (!this._adjacencyList.has(startVertex) || !this._adjacencyList.has(endVertex)) {
+    if (
+      !this._adjacencyList.has(startVertex) ||
+      !this._adjacencyList.has(endVertex)
+    ) {
       return 'INVALID NODE';
       // throw new Error('INVALID NODE');
     } else {
@@ -59,49 +49,6 @@ class Graph {
       console.log([vertex, edge]);
     }
   }
-
-  // Breadth-first: uses a queue to track where you've been.
-  breadth(startingNode) {
-    if (!startingNode) return 'INVALID NODE';
-    // Our queue class
-    const queue = new Queue();
-    // A set is an object that will only store unique keys.
-    const visitedNodes = new Set();
-
-    // let result = [];
-
-    queue.enqueue(startingNode);
-    visitedNodes.add(startingNode);
-    // result.push(startingNode.value);
-
-    while (!queue.isEmpty()) {
-      // Remove the first item from the queue.
-      const currentNode = queue.dequeue();
-      // Get the neighbors of the current node.
-      const neighbors = this.getNeighbors(currentNode);
-      // Loop over all the neighbors.
-      for (let neighbor of neighbors) {
-        const neighborNode = neighbor.vertex;
-        if (visitedNodes.has(neighborNode)) {
-          // If the Set has the node that I'm looking for, continue on to the next node.
-          continue;
-        } else {
-          // Otherwise, I haven't been there and I need to add it to the Set.
-          visitedNodes.add(neighborNode);
-          // Put it into the queue
-          queue.enqueue(neighborNode);
-          // result.push(neighborNode.value);
-        }
-      }
-    }
-    // console.log(result);
-    // Return the Set, so that there is a list of all visited nodes.
-    return visitedNodes;
-  }
 }
 
-module.exports = {
-  Vertex,
-  Edge,
-  Graph,
-};
+module.exports = Graph;
