+++ javascript/data-structure/graph/vertex.js
@@ -1,0 +1,9 @@
+'use strict';
+
+class Vertex {
+  constructor(value) {
+    this.value = value;
+  }
+}
+
+module.exports = Vertex;
